scripts/requirements.txt
python-docx
openai
